Conflicts:
	app/code/Magento/Catalog/Model/Indexer/Product/Price/Processor.php
	dev/tests/integration/testsuite/Magento/Catalog/Model/Indexer/Product/Flat/Action/RowTest.php
	dev/tests/unit/testsuite/Magento/Catalog/Model/Indexer/Product/Flat/ProcessorTest.php

+++ app/code/Magento/Reward/Controller/Adminhtml/Reward/Rate.php
@@ -31,10 +31,8 @@ class Rate extends \Magento\Backend\App\Action
      * @param \Magento\Backend\App\Action\Context $context
      * @param \Magento\Registry $coreRegistry
      */
-    public function __construct(
-        \Magento\Backend\App\Action\Context $context,
-        \Magento\Registry $coreRegistry
-    ) {
+    public function __construct(\Magento\Backend\App\Action\Context $context, \Magento\Registry $coreRegistry)
+    {
         $this->_coreRegistry = $coreRegistry;
         parent::__construct($context);
     }
@@ -47,8 +45,9 @@ class Rate extends \Magento\Backend\App\Action
      */
     public function dispatch(\Magento\App\RequestInterface $request)
     {
-        if (!$this->_objectManager->get('Magento\Reward\Helper\Data')->isEnabled()
-            && $request->getActionName() != 'noroute'
+        if (!$this->_objectManager->get(
+            'Magento\Reward\Helper\Data'
+        )->isEnabled() && $request->getActionName() != 'noroute'
         ) {
             $this->_forward('noroute');
         }
@@ -63,11 +62,15 @@ class Rate extends \Magento\Backend\App\Action
     protected function _initAction()
     {
         $this->_view->loadLayout();
-        $this->_setActiveMenu('Magento_Reward::customer_reward')
-            ->_addBreadcrumb(__('Customers'),
-                __('Customers'))
-            ->_addBreadcrumb(__('Manage Reward Exchange Rates'),
-                __('Manage Reward Exchange Rates'));
+        $this->_setActiveMenu(
+            'Magento_Reward::customer_reward'
+        )->_addBreadcrumb(
+            __('Customers'),
+            __('Customers')
+        )->_addBreadcrumb(
+            __('Manage Reward Exchange Rates'),
+            __('Manage Reward Exchange Rates')
+        );
         return $this;
     }
 
@@ -140,7 +143,7 @@ class Rate extends \Magento\Backend\App\Action
         if ($data) {
             $rate = $this->_initRate();
 
-            if ($this->getRequest()->getParam('rate_id') && ! $rate->getId()) {
+            if ($this->getRequest()->getParam('rate_id') && !$rate->getId()) {
                 return $this->_redirect('adminhtml/*/');
             }
 
@@ -189,36 +192,47 @@ class Rate extends \Magento\Backend\App\Action
     public function validateAction()
     {
         $response = new \Magento\Object(array('error' => false));
-        $post     = $this->getRequest()->getParam('rate');
-        $message  = null;
+        $post = $this->getRequest()->getParam('rate');
+        $message = null;
         /** @var \Magento\Core\Model\StoreManagerInterface $storeManager */
         $storeManager = $this->_objectManager->get('Magento\Core\Model\StoreManagerInterface');
         if ($storeManager->isSingleStoreMode()) {
             $post['website_id'] = $storeManager->getStore(true)->getWebsiteId();
         }
 
-        if (!isset($post['customer_group_id'])
-            || !isset($post['website_id'])
-            || !isset($post['direction'])
-            || !isset($post['value'])
-            || !isset($post['equal_value'])) {
+        if (!isset(
+            $post['customer_group_id']
+        ) || !isset(
+            $post['website_id']
+        ) || !isset(
+            $post['direction']
+        ) || !isset(
+            $post['value']
+        ) || !isset(
+            $post['equal_value']
+        )
+        ) {
             $message = __('Please enter all Rate information.');
-        } elseif ($post['direction'] == \Magento\Reward\Model\Reward\Rate::RATE_EXCHANGE_DIRECTION_TO_CURRENCY
-                  && ((int) $post['value'] <= 0 || (float) $post['equal_value'] <= 0)) {
-              if ((int) $post['value'] <= 0) {
-                  $message = __('Please enter a positive integer number in the left rate field.');
-              } else {
-                  $message = __('Please enter a positive number in the right rate field.');
-              }
-        } elseif ($post['direction'] == \Magento\Reward\Model\Reward\Rate::RATE_EXCHANGE_DIRECTION_TO_POINTS
-                  && ((float) $post['value'] <= 0 || (int) $post['equal_value'] <= 0)) {
-              if ((int) $post['equal_value'] <= 0) {
-                  $message = __('Please enter a positive integer number in the right rate field.');
-              } else {
-                  $message = __('Please enter a positive number in the left rate field.');
-              }
+        } elseif ($post['direction'] == \Magento\Reward\Model\Reward\Rate::RATE_EXCHANGE_DIRECTION_TO_CURRENCY &&
+            ((int)$post['value'] <= 0 ||
+            (double)$post['equal_value'] <= 0)
+        ) {
+            if ((int)$post['value'] <= 0) {
+                $message = __('Please enter a positive integer number in the left rate field.');
+            } else {
+                $message = __('Please enter a positive number in the right rate field.');
+            }
+        } elseif ($post['direction'] == \Magento\Reward\Model\Reward\Rate::RATE_EXCHANGE_DIRECTION_TO_POINTS &&
+            ((double)$post['value'] <= 0 ||
+            (int)$post['equal_value'] <= 0)
+        ) {
+            if ((int)$post['equal_value'] <= 0) {
+                $message = __('Please enter a positive integer number in the right rate field.');
+            } else {
+                $message = __('Please enter a positive number in the left rate field.');
+            }
         } else {
-            $rate       = $this->_initRate();
+            $rate = $this->_initRate();
             $isRateUnique = $rate->getIsRateUniqueToCurrent(
                 $post['website_id'],
                 $post['customer_group_id'],
@@ -226,7 +240,9 @@ class Rate extends \Magento\Backend\App\Action
             );
 
             if (!$isRateUnique) {
-                $message = __('Sorry, but a rate with the same website, customer group and direction or covering rate already exists.');
+                $message = __(
+                    'Sorry, but a rate with the same website, customer group and direction or covering rate already exists.'
+                );
             }
         }
